it's 1AM, why am I awake?
well, now the reports tab works as planned, there is a switch statement in the manager controller fills the reports scene as needed, it is a one scene to many JSON commands thingy, it works and it's efficient. (I think, idk, it's better than creating the same scene for multiple commands.) don't uncomment the cases that are commented until the api is complete, or do, I'm a commit message not a cop.