thing work good now
deleted the servicetickets menu item because it was redundant, now there is a products menu to update product amounts, or add a new product to the DB. tweeked some things with the table views in the main menu, if you select an item in one menu, any item in the other menu is deselected, also the edit current ticket button is disabled until a menu item is selected

crated the fxml needed for the update tickets and created the controller class(not yet implemented.) for the fxml.

if anyone can help me figure out how to import the sleep library into my life that would be great!